[#177058904] refactoring
